backend/foodgram/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.5
Django==3.2.16
django-debug-toolbar==2.2
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
Faker==12.0.1
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.2.0
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
pep8-naming==0.12.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2022.2.1
requests==2.26.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.3.2.post1
sqlparse==0.4.2
testfixtures==6.18.5
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.11
wcwidth==0.1.8
zipp==2.2.0
